Tweak the default color scheme to improve dark-mode contrast.
Restore `algue:14` as default, use `fira` where condensed is preferred.

Darken default `text-color` a bit and brighten in the mouseover condition.

`AG_Scrollbar`: Honor `background-color`. Improve contrast between the controls.

Move the `AG_Window` section on top of `style.css` since it's effectively the
root styleable class.



git-svn-id: https://dev.csoft.net/agar/trunk@11231 86a84d31-9702-0410-8f1e-a60cab7bd424

+++ sg/sg_script.c
@@ -1484,7 +1484,7 @@ Render(AG_Event *_Nonnull event)
 					continue;
 				}
 				for (c = &dent[0]; *c != '.' && *c != '\0'; c++) {
-					if (!isdigit(*c))
+					if (!isdigit((int)*c))
 						break;
 				}
 				if (*c == '.' || *c == '\0') {
